requirements.txt
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
feedparser==5.2.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
uWSGI==2.0.19.1
webencodings==0.5.1
Werkzeug==1.0.1
